refactor(sale): SaleEditDialog 3-way split (router + edit form + locked view)
지금까지 SaleEditDialog 한 파일이 라우팅(7일 lock 판정) + 편집 폼 +
잠금 뷰를 모두 책임. 본질적으로 분리된 두 화면이라 파일도 분리.

- SaleEditDialog.tsx: lock 분기만 → 두 view로 위임 (~15줄)
- SaleEditForm.tsx: 편집 폼 본체 (quantities/reason/저장/삭제)
- SaleLockedView.tsx: 7일 경과 read-only 요약

confirm() → AlertDialog 마이그레이션은 디자인 시스템 컴포넌트 추가가
선행되어야 해서 후속 PR로 분리 (스코프 가드).
